Use overloaded `pow` instead of `powf` to avoid non-compliance issues in GCC 11.4.0.  This will need to be changed back to `powf` for C++26 in the future.

+++ interface/src/raypick/PathPointer.cpp
@@ -340,7 +340,7 @@ void StartEndRenderState::update(const glm::vec3& origin, const glm::vec3& end, 
         glm::vec3 dim;
         if (distanceScaleEnd) {
             glm::vec3 camPos = qApp->getCamera().getPosition();
-            dim = getEndDim() * std::powf(glm::distance(camPos, end), 0.75f);
+            dim = getEndDim() * std::pow(glm::distance(camPos, end), 0.75f);
         } else {
             dim = getEndDim() * parentScale;
         }
